+++ src/components/About/AboutFooter.js
@@ -1,0 +1,49 @@
+/*
+Purpose
+  The About tabs, used in larger screens
+
+Child Components
+  - Not sure yet
+
+Libs
+  - Not sure yet
+
+API
+  - Not sure yets
+
+State needed
+  - Not sure yet
+
+Props
+  - Not sure yet
+*/
+import * as React from 'react';
+
+import Box from '@mui/material/Box';
+import Grid from '@mui/material/Grid';
+
+import { makeStyles } from '@mui/styles';
+
+import NEMACAboutLogoImage from '../../assets/images/nemac_logo_white.png';
+import NFWFAboutLogoImage from '../../assets/images/nfwf_logo_white.png';
+
+const useStyles = makeStyles((theme) => ({
+  AboutLogoImageStyle: {
+    maxHeight: '64px',
+  },
+}));
+
+export default function AboutFooter (props) {
+  const classes = useStyles();
+
+  return (
+    <Grid container spacing={2} justifyContent="center" alignItems="center" px={{ xs: 2, md: 3 }} pt={2} pb={0.75} >
+      <Grid item xs={12} sm={6} >
+        <img src={NFWFAboutLogoImage} className={classes.AboutLogoImageStyle} />
+      </Grid>
+      <Grid item xs={12} sm={6} >
+        <img src={NEMACAboutLogoImage} className={classes.AboutLogoImageStyle} />
+      </Grid>
+    </Grid>
+  )
+}
